HPm/eavesdrop/room-mic-and-remote-control/gui/requirements.txt
pygubu
pillow
numpy